♻️ : #38 - 모달 컴포넌트 분리작업 완료
